cleaned up a variety of small issues mostly related to tasks

+++ csharp/Hecatomb/Hecatomb/Components/Tasks/Task.cs
@@ -85,7 +85,7 @@ namespace Hecatomb
                 FetchIngredient();
                 return;
             }
-            if (Tiles.QuickDistance(Worker.X, Worker.Y, Worker.Z, Entity.X, Entity.Y, Entity.Z) <= WorkRange)
+            if ((int) Tiles.QuickDistance(Worker.X, Worker.Y, Worker.Z, Entity.X, Entity.Y, Entity.Z) <= WorkRange)
             {
                 Work();
             }
